+++ lib/src/widgets/chat/typing/typing_export.dart
@@ -1,2 +1,3 @@
-export 'ds_typing_message_bubble.widget.dart' show DSTypingMessageBubble;
+export 'ds_typing_message_bubble.widget.dart'
+    show DSTypingAnimationMessageBubble;
 export 'ds_dot_animation.widget.dart' show DSDotAnimation;
